sync

+++ app/utils/photo.py
@@ -5,14 +5,13 @@ import os
 
 from app.utils import mysql_util
 
-def upload_image(file):
+def upload_image(file: str):
     """
     Converts an image file to a png with a name
     matching its id in the database.
 
     file: file path string
     """
-    # handle both file paths and FileStorage objects
     if isinstance(file, str):  # if it's a file path
         original_path = file
         if not os.path.exists(file):
@@ -32,15 +31,24 @@ def upload_image(file):
     
     # save file as {photo_id}.png
     filepath = ('app/static/images/uploads/%s.png' % photo_id)
+    if (original_path == filepath):
+        return photo_id
     with open(filepath, 'wb') as f:
         f.write(png_buffer.getvalue())
 
-    # delete original
+    # delete original if different from original
     os.remove(original_path)
     
     return photo_id
 
+def link_item_photo(item_id: int, photo_id: int, display_order: int):
+    sql = '''
+    INSERT INTO Item_Photo (item_id, photo_id, display_order) VALUES
+    (%s, %s, %s);
+    '''
+    mysql_util.execute_sql(sql, (item_id, photo_id, display_order), True)
+
 if __name__ == '__main__':
     upload_image('app/static/images/uploads/harp.webp')
     upload_image('app/static/images/uploads/t1.png')
-    upload_image('app/static/images/uploads/t2.webp')
+    upload_image('app/static/images/uploads/t2.png')
